+++ app/routes/index.tsx
@@ -1,35 +1,4 @@
-import { Form } from "@remix-run/react";
-
-function Contact() {
-  return (
-    <Form action="https://submit-form.com/JZojw5Xb">
-      <input
-        style={inputStyle}
-        type="text"
-        placeholder="what's your name?"
-        name="name"
-      />
-      <input
-        style={inputStyle}
-        type="email"
-        placeholder="can I have your email?"
-        name="email"
-      />
-      <button
-        style={{
-          padding: "0.75rem 1.25rem",
-          borderRadius: 50,
-          border: "none",
-          color: "#1d1d1d",
-          background: "lightblue",
-        }}
-        type="submit"
-      >
-        Stay in touch
-      </button>
-    </Form>
-  );
-}
+import Contact from "~/components/contact";
 
 export default function Index() {
   return (
@@ -63,12 +32,3 @@ export default function Index() {
     </main>
   );
 }
-
-const inputStyle = {
-  display: "block",
-  marginBottom: 6,
-  padding: "0.75rem 0.5rem 0.75rem 1rem",
-  borderRadius: 50,
-  border: "1px solid #e2e2e2",
-  width: 320,
-};
